requirements.txt
pandas
sqlalchemy
chromadb
sentence-transformers
nltk
requests
python-dotenv